refactor shapes models
