+++ nanome/_internal/shapes/__init__.py
@@ -1,8 +1,3 @@
-from .shape import _Shape
-from .anchor import _Anchor
-from .line import _Line
-from .label import _Label
-from .sphere import _Sphere
-from .mesh import _Mesh
+from .models import _Shape, _Anchor, _Line, _Label, _Sphere, _Mesh
 from . import serializers
 from . import _deprecated
