--- nanome/_internal/shapes/label.py
@@ -1,15 +1,0 @@
-from .anchor import _Anchor
-from .shape import _Shape
-
-
-class _Label(_Shape):
-    def __init__(self):
-        from nanome.util.enums import ShapeType
-        _Shape.__init__(self, ShapeType.Label)
-        self._anchors = [_Anchor._create()]
-        self._text = ""
-        self._font_size = .5
-
-    @classmethod
-    def _create(cls):
-        return cls()
