--- nanome/_internal/shapes/sphere.py
@@ -1,14 +1,0 @@
-from .anchor import _Anchor
-from .shape import _Shape
-
-
-class _Sphere(_Shape):
-    def __init__(self):
-        from nanome.util.enums import ShapeType
-        _Shape.__init__(self, ShapeType.Sphere)
-        self._anchors = [_Anchor._create()]
-        self._radius = 1.0
-
-    @classmethod
-    def _create(cls):
-        return cls()
